2/26 실습 Application2 수정

+++ src/main/java/com/kkkwid/generic/level01/normal/Application2.java
@@ -5,7 +5,7 @@ public class Application2 {
 
     public static void main(String[] args) {
         Integer[] intArray = {1, 2, 3, 4, 5};
-        System.out.println("Integer 합계 : " + sum(intArray));
+        System.out.println("Integer 합계 : " + (int)sum(intArray));
 
         Double[] doubleArray = {1.1, 2.2, 3.3};
         System.out.println("Double 합계 : " + (double) sum(doubleArray));
@@ -14,10 +14,10 @@ public class Application2 {
 //         System.out.println("String 합계 : " + sum(strArray));  // 컴파일 에러 발생
     }
 
-    public static <T extends Number> int sum(T[] a){
-        int sum = 0;
+    public static <T extends Number> double sum(T[] a){
+        double sum = 0;
         for (T t : a) {
-            sum += t.intValue();
+            sum += t.doubleValue();
         }
         return sum;
     }
